pyproject.toml
[tool.poetry]
name = "iscc-resolver"
version = "0.1.0"
description = ""
authors = ["Titusz Pan <titusz.pan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.1"
mcrpc = "^2.0.6"
django-model-utils = "^4.0.0"
iscc = "^1.0.5"
loguru = "^0.5.2"
uvarint = "^1.2.0"
pytz = "^2020.1"
django-public-admin = "^0.0.3"
psycopg2 = "^2.8.6"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
